requirements/requirements.txt
django==2.2.7
django-settings-export==1.2.1
django-widget-tweaks==1.4.5
django-datatables-view==1.19.1
Pillow==6.2.1
python-dateutil
gunicorn
django-multiselectfield
psycopg2-binary
django-ckeditor
xlrd
django-model-utils
django-braces
django-resized
django-cleanup
jsonfield
pandas
selenium
webdriver-manager
packaging
pymongo
django-crontab
eeweather
skforecast
scikit-learn
Werkzeug==2.0.1
pretty-html-table
tqdm
